Handle API returning error due to no addresses found.
The blockchair API will return a 404 when none of the addresses
being queried are present in the blockchain database.  For the
report, in this case we still want to show the information with
0 payments made as it is the initial state.

In case we ever get an error code from blockchair API
we'll display the error message under the report.